Integrate PhysicsWorld to index and increase charm size and wall
