+++ src/components/physics/Charm.tsx
@@ -3,18 +3,19 @@ import { View, Image } from 'react-native';
 import { CharmPosition } from '@/src/types/charms.types';
 
 interface CharmProps {
+  radius: number;
   position: CharmPosition;
   source: any; // ImageSourcePropType would be more precise, but requires additional import
 }
 
-const Charm = memo(({ position, source }: CharmProps) => {
+const Charm = memo(({ radius, position, source }: CharmProps) => {
   return (
     <View
       style={[
         {
           position: 'absolute',
-          width: 40,
-          height: 40,
+          width: radius * 2,
+          height: radius * 2,
           left: position.x - 20,
           top: position.y - 20,
           transform: [{ rotate: `${position.angle}rad` }],
