+++ src/components/physics/PhysicsWorld.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { Dimensions } from 'react-native';
+import { Dimensions, StatusBar } from 'react-native';
 import Matter from 'matter-js';
 import { GameEngine } from 'react-native-game-engine';
 import { usePhysicsContext } from '../../contexts/PhysicsContext';
@@ -8,7 +8,12 @@ import Charm from './Charm';
 import { usePhysicsAnimation } from '@/src/hooks/usePhysicsAnimation';
 import { mockCharms } from '@/src/constants/constants';
 
-const { width, height } = Dimensions.get('window');
+// Get screen dimensions
+const { width, height: screenHeight } = Dimensions.get('window');
+// Account for status bar height
+const statusBarHeight = StatusBar.currentHeight || 0;
+// Usable height (accounting for status bar and some bottom padding for UI)
+const height = screenHeight - statusBarHeight - 200; // 180px bottom padding for UI elements
 
 interface PhysicsWorldProps {
   charms: Matter.Body[];
@@ -18,16 +23,57 @@ export const PhysicsWorld: React.FC<PhysicsWorldProps> = ({ charms }) => {
   const { engine } = usePhysicsContext();
   const positions = usePhysicsAnimation(engine, charms);
 
+  // Set up walls with proper positioning
   useEffect(() => {
-    // Create boundaries
+    // Wall thickness
+    const WALL_THICKNESS = 40;
+
+    // Create walls to form a bounded container
     const walls = [
-      Matter.Bodies.rectangle(width / 2, -30, width, 60, { isStatic: true }),
-      Matter.Bodies.rectangle(width / 2, height + 30, width, 60, { isStatic: true }),
-      Matter.Bodies.rectangle(-30, height / 2, 60, height, { isStatic: true }),
-      Matter.Bodies.rectangle(width + 30, height / 2, 60, height, { isStatic: true })
+      // Top wall - centered horizontally, positioned at top
+      Matter.Bodies.rectangle(
+        width / 2,
+        statusBarHeight,
+        width,
+        WALL_THICKNESS,
+        { isStatic: true }
+      ),
+      // Bottom wall - centered horizontally, positioned at bottom
+      Matter.Bodies.rectangle(
+        width / 2,
+        height + statusBarHeight,
+        width,
+        WALL_THICKNESS,
+        { isStatic: true }
+      ),
+      // Left wall - positioned at left edge + half thickness
+      Matter.Bodies.rectangle(
+        0,  // Changed from 0 to WALL_THICKNESS/2
+        height / 2 + statusBarHeight,
+        WALL_THICKNESS,
+        height,
+        { isStatic: true }
+      ),
+      // Right wall - positioned at right edge - half thickness
+      Matter.Bodies.rectangle(
+        width - WALL_THICKNESS / 2,  // Changed from width to width - WALL_THICKNESS/2
+        height / 2 + statusBarHeight,
+        WALL_THICKNESS,
+        height,
+        { isStatic: true }
+      ),
     ];
+
+    // Add walls to the world
     Matter.World.add(engine.world, walls);
 
+    // Adjust world bounds
+    engine.world.bounds = {
+      min: { x: 0, y: statusBarHeight },
+      max: { x: width, y: height + statusBarHeight }
+    };
+
+    // Cleanup function
     return () => {
       Matter.World.clear(engine.world, false);
     };
@@ -47,6 +93,7 @@ export const PhysicsWorld: React.FC<PhysicsWorldProps> = ({ charms }) => {
         <Charm
           source={mockCharms[index % mockCharms.length].image}
           key={charm.id}
+          radius={charm.circleRadius || 0}
           position={positions[charm.id] || {
             x: charm.position.x,
             y: charm.position.y,
